+++ src/components/AddBookForm.js
@@ -1,9 +1,9 @@
 /* eslint-disable camelcase */
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import axios from 'axios';
 import { useDispatch } from 'react-redux';
 import { v4 as uuid_v4 } from 'uuid';
-import { addBook } from '../redux/books/books';
+import { addBook, getBooks } from '../redux/books/books';
 import { baseUrl, appId } from '../redux/api';
 
 export const AddBookForm = () => {
@@ -11,11 +11,21 @@ export const AddBookForm = () => {
   const [author, setAuthor] = useState('');
   const dispatch = useDispatch();
 
+  const getBooksFromApi = () => {
+    // dispatch(getBooks);
+    axios.get(`${baseUrl}${appId}/books`).then((response) => {
+      if (response.data) {
+        dispatch(getBooks(response.data));
+      }
+    });
+  };
+
   const postBook = (newBook) => {
     axios.post(`${baseUrl}${appId}/books`, newBook)
       .then((response) => {
         if (response.status === 201) {
           dispatch(addBook(newBook));
+          getBooksFromApi();
         }
       })
       .catch((error) => { throw new Error(error); });
@@ -34,6 +44,10 @@ export const AddBookForm = () => {
     setTitle('');
     setAuthor('');
   };
+
+  useEffect(() => {
+    getBooksFromApi();
+  }, []);
 
   return (
     <div>
