+++ src/components/BookList.js
@@ -4,7 +4,7 @@ import { Book } from './Book';
 export const BookList = (prop) => ( 
   <div>
     <ul>
-      {prop.booksArray.map((data, index) => <Book key={index} book={data}/>)}
+      {prop && prop.booksArray.map((data, index) => <Book key={index} book={data}/>)}
     </ul>
     <br />
     <AddBookForm />
